feat(RHINENG-26251): split move system component
This commit splits the row action component from the helper functions
and improves the documentation.
